Add models module
